move room assignment to client side because firebase transactions are amazing
